rough new state management implementation
