+++ lib/plugin_api.dart
@@ -1,6 +1,7 @@
 library flutter_map.plugin_api;
 
 export 'package:flutter_map/flutter_map.dart';
+export 'package:flutter_map/src/map/flutter_map_state.dart';
 export 'package:flutter_map/src/core/bounds.dart';
 export 'package:flutter_map/src/core/center_zoom.dart';
 export 'package:flutter_map/src/map/map.dart';
