+++ lib/src/layer/tile_layer/transformation_calculator.dart
@@ -1,13 +1,13 @@
 import 'package:flutter_map/src/layer/tile_layer/level.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_transformation.dart';
-import 'package:flutter_map/src/map/map.dart';
+import 'package:flutter_map/src/map/flutter_map_state.dart';
 
 class TransformationCalculator {
   final Map<double, Level> _levels = {};
 
   Level? levelAt(double zoom) => _levels[zoom];
 
-  Level getOrCreateLevel(double zoom, MapState map) {
+  Level getOrCreateLevel(double zoom, FlutterMapState map) {
     final level = _levels[zoom];
     if (level != null) return level;
 
@@ -30,7 +30,7 @@ class TransformationCalculator {
     _levels.remove(levelZoom);
   }
 
-  TileTransformation transformationFor(double levelZoom, MapState map) {
+  TileTransformation transformationFor(double levelZoom, FlutterMapState map) {
     final level = _levels[levelZoom]!;
     final scale = map.getZoomScale(map.zoom, level.zoom);
     final pixelOrigin = map.getNewPixelOrigin(map.center, map.zoom).round();
